extracted common compiler logic
